ccp_marketing/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ccp-marketing"
version = "0.1.0"
description = "CCP Digital Marketing - Event creation and social media promotion using Composio SDK"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Coffee Code Philly" }
]
keywords = ["marketing", "events", "social-media", "automation", "composio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
    "Topic :: Office/Business",
]

dependencies = [
    "composio>=0.7.0",
    "typer>=0.9.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
ccp-marketing = "ccp_marketing.cli.main:app"

[project.urls]
Homepage = "https://github.com/coffeecodePhilly/ccp-marketing"
Documentation = "https://github.com/coffeecodePhilly/ccp-marketing#readme"
Issues = "https://github.com/coffeecodePhilly/ccp-marketing/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ccp_marketing"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src/ccp_marketing --cov-report=term-missing"
